backend/requirements.txt
fastapi>=0.115.0,<1.0
uvicorn[standard]>=0.32.0,<1.0
pydantic>=2.10.0,<3.0
python-dotenv>=1.0.0,<2.0
python-multipart>=0.0.18,<1.0
mediapipe>=0.10.21,<1.0
gtts>=2.5.0,<3.0
websockets>=14.0,<15.0
Pillow>=10.0.0,<12.0
numpy>=1.26.0,<3.0
openai>=1.50.0,<2.0

# Dev dependencies
# ruff>=0.9.0  # linter/formatter — install with: pip install ruff
